Fix eslint no-cycle error and input focus on edit
